+++ spec/factories/pipelines.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 FactoryGirl.define do
   factory :pipeline do
-    name { Faker::Name.name }
+    name { Faker::Name.unique.name }
     patient
     user
   end